==> xgboost/non_imputed/5_fairness_assess_get_data_xgboost.ipynb <==
# Data for fairness assessment
Get the columns of data we want to assess
Save for each imputation set

              inc_age = age
              sex
              dialtyp=type of dialysis
              race

import psycopg2
from sqlalchemy import create_engine

# other libraries
import numpy as np
import pandas as pd
import sys
import pickle

print('python-' + sys.version)

con = create_engine('postgresql://username:password@location/dbname')

#import only the columns we want to look at in this assessment
df = pd.read_sql_query('''SELECT   usrds_id, died_in_90, inc_age, sex, dialtyp, race, hispanic, subset FROM medxpreesrd;''', con)
df.head()

#save the dataset
with open('../complete_fairness_data.pickle', 'wb') as picklefile:  
    pickle.dump(df, picklefile)
