==> xgboost/non_imputed/6_xgb_nonimputed_fairness.ipynb <==
   # Fairness

import numpy as np
import pandas as pd
import pickle
 
import datetime
dte = datetime.datetime.now()
dte = dte.strftime("%Y%m%d")

# import custom functions
import sys
sys.path.append('../../onc_functions')

from fairness import get_fairness_assessment

print('python-' + sys.version)

# Calculate Fairness Assessment
- get the true y-values and the y_prediction_probability from the model
- get details to evaluate for fairness

# load results from the model
pred_df = pd.read_csv('./roc_auc/2021_xgb_nonimputed_y_proba.csv')
pred_df.drop(columns=['Unnamed: 0'],inplace=True)
pred_df.head()

with open('../complete_fairness_data.pickle', 'rb') as f:
        dataset = pickle.load(f)
dataset.head()

# merge results with fairness details
data = pred_df.merge(dataset, how='left', on=['usrds_id'])
data.head()

fairness = get_fairness_assessment(
                                        data,
                                        y_proba_col_name='score',
                                        y_true_col_name='died_in_90')
fairness.head(39)

#save (for the aucroc evaluated model)
fairness.to_csv('./roc_auc/' + str(dte) + '_xgb_nonimputed_fairness.csv')
